+++ public_transport_watcher/extractor/configuration/extraction_config.py
@@ -13,4 +13,16 @@ EXTRACTION_CONFIG = {
     "addresses": {
         "batch_size": 1000,
     },
+    "pollution": {
+        "pollutants": ["NO2", "O3", "PM2.5"],
+        "limits": {
+            # See https://www.airparif.fr/la-reglementation-en-france
+            "NO2": 40,
+            "O3": 120,
+            "PM2.5": 25,
+        },
+    },
+    "weather": {
+        "station": "ME099",
+    },
 }