+++ public_transport_watcher/extractor/extract/__init__.py
@@ -1,11 +1,15 @@
 from .addresses import extract_addresses_informations
+from .air_quality import extract_air_quality_data
 from .navigo import extract_navigo_validations
 from .stations import extract_stations_informations
 from .traffic import process_traffic_data
+from .weather import extract_weather_data
 
 __all__ = [
     "extract_addresses_informations",
+    "extract_air_quality_data",
     "extract_navigo_validations",
     "extract_stations_informations",
     "process_traffic_data",
+    "extract_weather_data",
 ]