+++ public_transport_watcher/extractor/extractor.py
@@ -1,13 +1,17 @@
 from public_transport_watcher.extractor.configuration import EXTRACTION_CONFIG
 from public_transport_watcher.extractor.extract import (
     extract_addresses_informations,
+    extract_air_quality_data,
     extract_navigo_validations,
     extract_stations_informations,
+    extract_weather_data,
     process_traffic_data,
 )
-from public_transport_watcher.utils import get_query_result
-
+from public_transport_watcher.extractor.extract.real_time import (
+    get_latest_air_quality_csv,
+)
 from public_transport_watcher.logging_config import get_logger
+from public_transport_watcher.utils import get_query_result
 
 logger = get_logger()
 
@@ -37,6 +41,17 @@ class Extractor:
     def extract_traffic_data(self):
         return process_traffic_data()
 
+    def extract_air_quality_data(self):
+        get_latest_air_quality_csv()
+        config = self.extract_config.get("pollution", {})
+        pollutants = config.get("pollutants", [])
+        limits = config.get("limits", {})
+        return extract_air_quality_data(pollutants, limits)
+
+    def extract_weather_data(self):
+        config = self.extract_config.get("weather", {})
+        return extract_weather_data(config)
+
 
 if __name__ == "__main__":
     extractor = Extractor()
@@ -44,3 +59,5 @@ if __name__ == "__main__":
     extractor.extract_navigo_validations()
     extractor.extract_addresses_informations()
     traffic_data = extractor.extract_traffic_data()
+    air_quality_data = extractor.extract_air_quality_data()  # needs to be scheduled
+    weather_data = extractor.extract_weather_data()  # needs to be scheduled