+++ public_transport_watcher/utils/__init__.py
@@ -1,13 +1,15 @@
-from .get_engine import get_engine
 from .get_credentials import get_credentials
 from .get_datalake_file import get_datalake_file
+from .get_engine import get_engine
+from .get_env_variable import get_env_variable
 from .get_query_result import get_query_result
-
+from .get_trafic_api import fetch_generic_api_data
 
 __all__ = [
-    "get_engine",
     "get_credentials",
     "get_datalake_file",
+    "get_engine",
+    "get_env_variable",
     "get_query_result",
-    "get_trafic_api",
+    "fetch_generic_api_data",
 ]